21: Moving everything under the io.github.turtlemonvh namespace to prepare for publishing to sonatype
