+++ src/main/java/io/github/turtlemonvh/ionicsparkutils/KeyServicesMinimal.java
@@ -1,4 +1,4 @@
-package com.ionic.sparkutil;
+package io.github.turtlemonvh.ionicsparkutils;
 
 import com.ionic.sdk.key.KeyServices;
 import com.ionic.sdk.agent.request.createkey.CreateKeysResponse;
