+++ src/main/java/io/github/turtlemonvh/ionicsparkutils/SDeviceProfile.java
@@ -1,4 +1,4 @@
-package com.ionic.sparkutil;
+package io.github.turtlemonvh.ionicsparkutils;
 
 import com.ionic.sdk.device.profile.DeviceProfile;
