+++ src/main/java/io/github/turtlemonvh/ionicsparkutils/TestAgent.java
@@ -1,4 +1,4 @@
-package com.ionic.sparkutil;
+package io.github.turtlemonvh.ionicsparkutils;
 
 /*
