+++ src/test/scala/io/github/turtlemonvh/ionicsparkutils/CredentialsTest.scala
@@ -1,4 +1,4 @@
-package com.ionic.sparkutil
+package io.github.turtlemonvh.ionicsparkutils
 
 /**
  * A simple test for for working with credentials.
